requirements.txt
# ----------------------
# Flask
# ----------------------
Flask
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-bcrypt
requests
psycopg2
pyjwt
#-------------------------
click==6.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
WTForms==2.2.1

# ----------------------
# Production Server
# ----------------------
gunicorn
